Types, refactoring, and cleanup
