+++ ts_src/speedrun/resources.ts
@@ -15,7 +15,7 @@ export class Variable implements IVariable {
         this.values = v ? mapValueArray(v.values.values) : null;
     }
 
-    @TryCatch({log: true})
+    @TryCatch()
     getValue(label: string) {
         return this.values.find(v => v.label === label);
     }
@@ -36,7 +36,7 @@ export class Category implements ICategory {
         this.variables = c ? c.variables.data.map((v: any) => new Variable(v)) : null;
     }
 
-    @TryCatch({log: true})
+    @TryCatch()
     getVariableIDs(name: string, value: string) {
         let ret = this.variables.find(v => v.name === name);
         return ret ? [ret.id, ret.getValue(value).id] : null;
@@ -60,7 +60,7 @@ export class Level implements ILevel {
         this.variables = l.variables ? l.variables.data.map((v: any) => new Variable(v)) : null;
     }
 
-    @TryCatch({log: true})
+    @TryCatch()
     getVariableIDs(name: string, value: string)
     {
         let ret = this.variables.find(v => v.name === name);
@@ -87,18 +87,18 @@ export class Game implements IGame {
         this.levels = g.levels ? g.levels.data.map(l => new Level(l)) : null;
     }
 
-    @TryCatch({ log: true })
+    @TryCatch()
     getVariableIDs(name: string, value: string) {
         let ret = this.variables.find(v => v.name === name);
         return ret ? [ret.id, ret.getValue(value).id] : null;
     }
 
-    @TryCatch({ log: true })
+    @TryCatch()
     getCategory(name: string) {
         return this.categories.find(c => c.name === name);
     }
 
-    @TryCatch({ log: true })
+    @TryCatch()
     getLevel(name: string) {
         return this.levels.find(l => l.name === name);
     }
