+++ ts_src/utils/types.ts
@@ -1,0 +1,4 @@
+
+export type fn = Function;
+export type bool = boolean;
+export type Decorator<T> = (target: T, key?: PropertyKey, method?: PropertyDescriptor) => any;
